src/ingestion/requirements.txt
requests
confluent-kafka
pandas
kafka-python
pymongo
python-dotenv
